requirements.txt
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.4
django-cors-headers==4.2.0
django-filter==23.2
django-otp==1.2.2
django-phonenumber-field==7.1.0
django-rest-framework==0.1.0
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.4
frozenlist==1.4.0
idna==3.4
inflection==0.5.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
Markdown==3.4.4
multidict==6.0.4
Pillow==10.0.0
psycopg2-binary==2.9.7
PyJWT==2.8.0
pyotp==2.9.0
python-http-client==3.3.7
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.2
sendgrid==6.10.0
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
twilio==8.7.0
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.4
yarl==1.9.2
